requirements.txt
numpy
pandas
scikit-learn

Pillow
matplotlib
opencv-python

tensorboard
pyyaml

mediapipe